+++ deepchem/models/tests/test_nnxc.py
@@ -26,28 +26,56 @@ def dummymodel():
 
 
 @pytest.mark.dqc
-def test_nnlda():
+def test_nnlda_cuda():
     torch.manual_seed(42)
     # https://github.com/diffqc/dqc/blob/742eb2576418464609f942def4fb7c3bbdc0cd82/dqc/test/test_xc.py#L15
     n = 2
     model = dummymodel()
-    k = NNLDA(model, device="cuda:0")
+    k = NNLDA(model, device="cuda")
     densinfo = ValGrad(
         value=torch.rand((n,), dtype=torch.float32).requires_grad_())
     output = k.get_edensityxc(densinfo)
-    expected_output = torch.tensor([0.3386, 0.0177]).to("cuda:0")
+    expected_output = torch.tensor([0.3386, 0.0177]).to("cuda")
     torch.testing.assert_close(output, expected_output, atol=1e-4, rtol=0)
 
 
 @pytest.mark.dqc
-def test_hybridxc():
+def test_hybridxc_cuda():
     torch.manual_seed(42)
     n = 2
     nnmodel = dummymodel()
-    k = HybridXC("lda_x", nnmodel, aweight0=0.0, device="cuda:0")
+    k = HybridXC("lda_x", nnmodel, aweight0=0.0, device="cuda")
     densinfo = ValGrad(
         value=torch.rand((n,), dtype=torch.float32).requires_grad_())
     output = k.get_edensityxc(densinfo)
     expected_output = torch.tensor([-0.6988, -0.2108],
-                                   dtype=torch.float64).to("cuda:0")
+                                   dtype=torch.float64).to("cuda")
+    torch.testing.assert_close(output, expected_output, atol=1e-4, rtol=0)
+
+
+@pytest.mark.dqc
+def test_nnlda_cpu():
+    torch.manual_seed(42)
+    # https://github.com/diffqc/dqc/blob/742eb2576418464609f942def4fb7c3bbdc0cd82/dqc/test/test_xc.py#L15
+    n = 2
+    model = dummymodel()
+    k = NNLDA(model, device="cpu")
+    densinfo = ValGrad(
+        value=torch.rand((n,), dtype=torch.float32).requires_grad_())
+    output = k.get_edensityxc(densinfo)
+    expected_output = torch.tensor([0.3386, 0.0177]).to("cpu")
+    torch.testing.assert_close(output, expected_output, atol=1e-4, rtol=0)
+
+
+@pytest.mark.dqc
+def test_hybridxc_cpu():
+    torch.manual_seed(42)
+    n = 2
+    nnmodel = dummymodel()
+    k = HybridXC("lda_x", nnmodel, aweight0=0.0, device="cpu")
+    densinfo = ValGrad(
+        value=torch.rand((n,), dtype=torch.float32).requires_grad_())
+    output = k.get_edensityxc(densinfo)
+    expected_output = torch.tensor([-0.6988, -0.2108],
+                                   dtype=torch.float64).to("cpu")
     torch.testing.assert_close(output, expected_output, atol=1e-4, rtol=0)
